# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: A3 Sales OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get all tasks
      parameters:
        - name: leadId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: dueFilter
          in: query
          schema:
            type: string
            enum: [today, overdue, next7days]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tasks/{id}/complete:
    patch:
      operationId: completeTask
      tags: [tasks]
      summary: Mark task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /templates:
    get:
      operationId: getTemplates
      tags: [templates]
      summary: Get all templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a new template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{id}:
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Task:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        taskType:
          type: string
        dueDate:
          type: string
        status:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        leadCompanyName:
          type: string
        leadContactName:
          type: string
      required:
        - id
        - taskType
        - status
        - createdAt
        - updatedAt
    TaskInput:
      type: object
      properties:
        leadId:
          type: integer
        taskType:
          type: string
        dueDate:
          type: string
        status:
          type: string
        notes:
          type: string
      required:
        - taskType
        - status
    Template:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        subject:
          type: string
        body:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - body
        - createdAt
        - updatedAt
    TemplateInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        subject:
          type: string
        body:
          type: string
      required:
        - name
        - category
        - body
...