# lib/api-zod/src/generated/types/templateInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A3 Sales OS API
 * OpenAPI spec version: 0.1.0
 */

export interface TemplateInput {
  name: string;
  category: string;
  subject?: string;
  body: string;
}